+++ account-api/src/main/java/com/example/accountapi/web/controller/account/AccountInfoController.java
@@ -8,8 +8,8 @@ import com.example.accountapi.domain.model.Customer;
 import com.example.accountapi.domain.model.Seller;
 import com.example.accountapi.web.validation.exception.AccountException;
 import com.example.accountapi.web.validation.exception.ErrorCode;
-import com.example.accountapi.config.SpringSecurity.CustomerPrincipalDetails;
-import com.example.accountapi.config.SpringSecurity.SellerPrincipalDetails;
+import com.example.accountapi.config.SpringSecurity.customer.CustomerPrincipalDetails;
+import com.example.accountapi.config.SpringSecurity.seller.SellerPrincipalDetails;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
