+++ account-api/src/main/java/com/example/accountapi/web/validation/form/LogInForm.java
@@ -1,0 +1,10 @@
+package com.example.accountapi.web.validation.form;
+
+import lombok.Data;
+
+@Data
+public class LogInForm {
+
+    private String email;
+    private String password;
+}
